// src/app/actions/upsert-appointment/schema.ts
import { z } from "zod";

export const upsertAppointmentSchema = z.object({
  id: z.string().uuid().optional(),
  patientId: z.string().uuid({ message: "Paciente é obrigatório" }),
  doctorId: z.string().uuid({ message: "Médico é obrigatório" }),
  appointmentPriceInCents: z
    .number()
    .min(1, { message: "Valor é obrigatório" }),
  date: z.date({ message: "Data é obrigatória" }),
  time: z.string().min(1, { message: "Horário é obrigatório" }),
});

export type UpsertAppointmentSchema = z.infer<typeof upsertAppointmentSchema>;